+++ src/main/java/myRetail/repository/ProductRepository.java
@@ -1,5 +1,6 @@
 package myRetail.repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import myRetail.model.Product;
@@ -10,5 +11,5 @@ public interface ProductRepository extends MongoRepository<Product, String> {
     public Product findByName(String firstName);
     public Product findBySequence(Integer firstName);
     public List<Product> findByPrice(String lastName);
-    public Product findById(Integer id);
+    public Product findById(BigInteger id);
 }
